requirements.txt
anyio==3.7.0
apache-libcloud==3.7.0
asn1crypto==1.5.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
debugpy==1.6.7
fastapi==0.97.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
pg8000==1.29.6
pydantic==1.10.9
pydub==0.25.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
scramp==1.4.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.16
sqlalchemy-file==0.4.0
starlette==0.27.0
typing_extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0
